+++ pages/system/app/list.vue
@@ -3,9 +3,12 @@
 		<view class="uni-header">
 			<uni-stat-breadcrumb class="uni-stat-breadcrumb-on-phone" />
 			<view class="uni-group">
-				<input class="uni-search" type="text" v-model="query" @confirm="search" :placeholder="$t('common.placeholder.query')" />
-				<button class="uni-button hide-on-phone" type="default" size="mini" @click="search">{{$t('common.button.search')}}</button>
-				<button class="uni-button" type="primary" size="mini" @click="navigateTo('./add')">{{$t('common.button.add')}}</button>
+				<input class="uni-search" type="text" v-model="query" @confirm="search"
+					:placeholder="$t('common.placeholder.query')" />
+				<button class="uni-button hide-on-phone" type="default" size="mini"
+					@click="search">{{$t('common.button.search')}}</button>
+				<button class="uni-button" type="primary" size="mini"
+					@click="navigateTo('./add')">{{$t('common.button.add')}}</button>
 				<button class="uni-button" type="warn" size="mini" :disabled="!selectedIndexs.length"
 					@click="delTable">{{$t('common.button.batchDelete')}}</button>
 				<!-- #ifdef H5 -->
@@ -23,24 +26,26 @@
 				:where="where" page-data="replace" :orderby="orderby" :getcount="true" :page-size="options.pageSize"
 				:page-current="options.pageCurrent" v-slot:default="{data,pagination,loading,error,options}"
 				:options="options" loadtime="manual" @load="onqueryload">
-				<uni-table ref="table" :loading="loading || addAppidLoading" :emptyText="error.message || $t('common.empty')"
-					border stripe type="selection" @selection-change="selectionChange" class="table-pc">
+				<uni-table ref="table" :loading="loading || addAppidLoading"
+					:emptyText="error.message || $t('common.empty')" border stripe type="selection"
+					@selection-change="selectionChange" class="table-pc">
 					<uni-tr>
 						<uni-th align="center" filter-type="search" @filter-change="filterChange($event, 'appid')"
 							sortable @sort-change="sortChange($event, 'appid')">AppID</uni-th>
 						<uni-th align="center" filter-type="search" @filter-change="filterChange($event, 'name')"
 							sortable @sort-change="sortChange($event, 'name')">应用名称</uni-th>
-						<uni-th align="center" width="380" filter-type="search" @filter-change="filterChange($event, 'description')"
-							sortable @sort-change="sortChange($event, 'description')">应用描述</uni-th>
+						<uni-th align="center" filter-type="search" @filter-change="filterChange($event, 'description')"
+							sortable @sort-change="sortChange($event, 'description')" :width="descriptionThWidth">应用描述
+						</uni-th>
 						<uni-th align="center" filter-type="timestamp"
 							@filter-change="filterChange($event, 'create_date')" sortable
 							@sort-change="sortChange($event, 'create_date')">创建时间</uni-th>
-						<uni-th align="center">操作</uni-th>
+						<uni-th align="center" :width="buttonThWidth">操作</uni-th>
 					</uni-tr>
-					<uni-tr v-for="(item,index) in data" :key="index">
+					<uni-tr v-for="(item,index) in data" :key="index" :disabled="item.appid === appid">
 						<uni-td align="center">{{item.appid}}</uni-td>
 						<uni-td align="center">{{item.name}}</uni-td>
-						<uni-td align="center">{{item.description}}</uni-td>
+						<uni-td align="left">{{item.description}}</uni-td>
 						<uni-td align="center">
 							<uni-dateformat :threshold="[0, 0]" :date="item.create_date"></uni-dateformat>
 						</uni-td>
@@ -49,8 +54,14 @@
 								-
 							</view>
 							<view v-else class="uni-group">
-								<button @click="navigateTo('./edit?id='+item._id, false)" class="uni-button" size="mini"
-									type="primary">{{$t('common.button.edit	')}}</button>
+								<button @click="publish(item._id)" class="uni-button" size="mini"
+									type="primary">{{$t('common.button.publish')}}</button>
+								<button
+									@click="navigateTo('/uni_modules/uni-upgrade-center/pages/version/list?appid='+item.appid, false)"
+									class="uni-button" size="mini"
+									type="primary">{{$t('common.button.version')}}</button>
+								<button @click="navigateTo('./add?id='+item.appid, false)" class="uni-button"
+									size="mini" type="primary">{{$t('common.button.edit')}}</button>
 								<button @click="confirmDelete(item._id)" class="uni-button" size="mini"
 									type="warn">{{$t('common.button.delete')}}</button>
 							</view>
@@ -59,17 +70,8 @@
 				</uni-table>
 
 				<view class="uni-pagination-box">
-					<!-- #ifndef MP -->
-					<picker class="select-picker" mode="selector" :value="pageSizeIndex" :range="pageSizeOption"
-						@change="changeSize">
-						<button type="default" size="mini" :plain="true">
-							<text>{{pageSizeOption[pageSizeIndex]}} {{$t('common.piecePerPage')}}</text>
-							<uni-icons class="select-picker-icon" type="arrowdown" size="12" color="#999"></uni-icons>
-						</button>
-					</picker>
-					<!-- #endif -->
-					<uni-pagination show-icon :page-size="pagination.size" v-model="pagination.current"
-						:total="pagination.count" @change="onPageChanged" />
+					<uni-pagination show-icon show-page-size :page-size="pagination.size" v-model="pagination.current"
+						:total="pagination.count" @change="onPageChanged" @pageSizeChange="pageSizeChange" />
 				</view>
 			</unicloud-db>
 		</view>
@@ -79,7 +81,6 @@
 		<!-- #endif -->
 	</view>
 </template>
-
 <script>
 	import {
 		enumConverter,
@@ -91,7 +92,7 @@
 
 	const db = uniCloud.database()
 	// 表查询配置
-	const dbOrderBy = '' // 排序字段
+	const dbOrderBy = 'create_date' // 排序字段
 	const dbSearchFields = [] // 模糊搜索字段，支持模糊搜索的字段列表
 	// 分页配置
 	const pageSize = 20
@@ -110,8 +111,6 @@
 				orderby: dbOrderBy,
 				orderByFieldName: "",
 				selectedIndexs: [],
-				pageSizeIndex: 0,
-				pageSizeOption: [20, 50, 100, 500],
 				options: {
 					pageSize,
 					pageCurrent,
@@ -133,7 +132,9 @@
 					}
 				},
 				exportExcelData: [],
-				addAppidLoading: true
+				addAppidLoading: true,
+				descriptionThWidth: 380,
+				buttonThWidth: 400
 			}
 		},
 		onLoad() {
@@ -145,19 +146,14 @@
 		computed: {
 			...mapState('app', ['appName', 'appid'])
 		},
-		watch: {
-			pageSizeIndex: {
-				immediate: true,
-				handler(val, old) {
-					this.options.pageSize = this.pageSizeOption[val]
-					this.options.pageCurrent = 1
-					this.$nextTick(() => {
-						this.loadData()
-					})
-				}
-			}
-		},
 		methods: {
+			pageSizeChange(pageSize) {
+				this.options.pageSize = pageSize
+				this.options.pageCurrent = 1
+				this.$nextTick(() => {
+					this.loadData()
+				})
+			},
 			onqueryload(data) {
 				if (!data.find(item => item.appid === this.appid)) {
 					this.addCurrentAppid({
@@ -181,7 +177,7 @@
 					setTimeout(() => {
 						uni.showModal({
 							content: `检测到数据库中无当前应用, 已自动添加应用: ${this.appName}`,
-							showCancel:false
+							showCancel: false
 						})
 					}, 500)
 				}).catch((err) => {
@@ -233,7 +229,9 @@
 			},
 			// 批量删除
 			delTable() {
-				console.warn("删除应用，只能删除应用表 opendb-app-list 中的应用数据记录，不能删除与应用关联的其他数据，例如：使用升级中心 uni-upgrade-center 等插件产生的数据（应用版本数据等）")
+				console.warn(
+					"删除应用，只能删除应用表 opendb-app-list 中的应用数据记录，不能删除与应用关联的其他数据，例如：使用升级中心 uni-upgrade-center 等插件产生的数据（应用版本数据等）"
+				)
 				this.$refs.udb.remove(this.selectedItems(), {
 					success: (res) => {
 						this.$refs.table.clearSelection()
@@ -245,8 +243,11 @@
 				this.selectedIndexs = e.detail.index
 			},
 			confirmDelete(id) {
-				console.warn("删除应用，只能删除应用表 opendb-app-list 中的应用数据记录，不能删除与应用关联的其他数据，例如：使用升级中心 uni-upgrade-center 等插件产生的数据（应用版本数据等）")
+				console.warn(
+					"删除应用，只能删除应用表 opendb-app-list 中的应用数据记录，不能删除与应用关联的其他数据，例如：使用升级中心 uni-upgrade-center 等插件产生的数据（应用版本数据等）"
+				)
 				this.$refs.udb.remove(id, {
+					confirmContent: '是否删除该应用',
 					success: (res) => {
 						this.$refs.table.clearSelection()
 					}
@@ -277,6 +278,11 @@
 				}
 				this.$nextTick(() => {
 					this.$refs.udb.loadData()
+				})
+			},
+			publish(id) {
+				uni.navigateTo({
+					url: '/pages/system/app/uni-portal/uni-portal?id=' + id
 				})
 			}
 		}