+++ pages/uni-stat/device/activity/activity.vue
@@ -39,15 +39,9 @@
 			<view class="uni-stat--x p-m">
 				<uni-stat-table :data="tableData" :filedsMap="fieldsMap" :loading="loading" tooltip />
 				<view class="uni-pagination-box">
-					<picker class="select-picker" mode="selector" :value="options.pageSizeIndex"
-						:range="options.pageSizeRange" @change="changePageSize">
-						<button type="default" size="mini" :plain="true">
-							<text>{{pageSize}} 条/页</text>
-							<uni-icons class="select-picker-icon" type="arrowdown" size="12" color="#999"></uni-icons>
-						</button>
-					</picker>
-					<uni-pagination show-icon :page-size="pageSize" :current="options.pageCurrent"
-						:total="options.total" @change="changePageCurrent" />
+					<uni-pagination show-icon show-page-size :page-size="options.pageSize"
+						:current="options.pageCurrent" :total="options.total" @change="changePageCurrent"
+						@pageSizeChange="changePageSize" />
 				</view>
 			</view>
 		</view>
@@ -87,10 +81,9 @@
 					start_time: [],
 				},
 				options: {
+					pageSize: 20,
 					pageCurrent: 1, // 当前页
 					total: 0, // 数据总量
-					pageSizeIndex: 0, // 与 pageSizeRange 一起计算得出 pageSize
-					pageSizeRange: [10, 20, 50, 100],
 				},
 				loading: false,
 				currentDateTab: 0,
@@ -102,13 +95,6 @@
 			}
 		},
 		computed: {
-			pageSize() {
-				const {
-					pageSizeRange,
-					pageSizeIndex
-				} = this.options
-				return pageSizeRange[pageSizeIndex]
-			},
 			chartTabs() {
 				const tabs = [{
 					_id: 'day',
@@ -173,7 +159,7 @@
 				this.getChannelData(null, id)
 				this.query.version_id = 0
 				this.query.uni_platform = item.code
-				console.log('this.query.uni_platform = item.code',item.code);
+				console.log('this.query.uni_platform = item.code', item.code);
 			},
 			changeTimeRange(id, index) {
 				this.currentDateTab = index
@@ -192,12 +178,9 @@
 				this.getTabelData(this.query)
 			},
 
-			changePageSize(e) {
-				const {
-					value
-				} = e.detail
+			changePageSize(pageSize) {
+				this.options.pageSize = pageSize
 				this.options.pageCurrent = 1 // 重置分页
-				this.options.pageSizeIndex = value
 				this.getTabelData(this.query)
 			},
 
@@ -208,7 +191,7 @@
 			},
 
 			getAllData(query) {
-				this.getChartData(query, this.currentChartTab,this.tabName)
+				this.getChartData(query, this.currentChartTab, this.tabName)
 				this.getTabelData(query)
 			},
 
@@ -221,7 +204,7 @@
 						data: []
 					}]
 				}
-				query = stringifyQuery(query,false,['uni_platform'])
+				query = stringifyQuery(query, false, ['uni_platform'])
 				const db = uniCloud.database()
 				if (type === 'day') {
 					db.collection(this.tableName)
@@ -286,8 +269,8 @@
 					.groupBy('start_time')
 					.groupField(stringifyGroupField(fieldsMap, field))
 					.orderBy('start_time', 'desc')
-					.skip((pageCurrent - 1) * this.pageSize)
-					.limit(this.pageSize)
+					.skip((pageCurrent - 1) * this.options.pageSize)
+					.limit(this.options.pageSize)
 					.get({
 						getCount: true
 					})