+++ pages/uni-stat/user/trend/trend.vue
@@ -44,15 +44,9 @@
 			<view class="uni-stat--x p-m">
 				<uni-stat-table :data="tableData" :filedsMap="fieldsMap" :loading="loading" />
 				<view class="uni-pagination-box">
-					<picker class="select-picker" mode="selector" :value="options.pageSizeIndex"
-						:range="options.pageSizeRange" @change="changePageSize">
-						<button type="default" size="mini" :plain="true">
-							<text>{{pageSize}} 条/页</text>
-							<uni-icons class="select-picker-icon" type="arrowdown" size="12" color="#999"></uni-icons>
-						</button>
-					</picker>
-					<uni-pagination show-icon :page-size="pageSize" :current="options.pageCurrent"
-						:total="options.total" @change="changePageCurrent" />
+					<uni-pagination show-icon show-page-size :page-size="options.pageSize"
+						:current="options.pageCurrent" :total="options.total" @change="changePageCurrent"
+						@pageSizeChange="changePageSize" />
 				</view>
 			</view>
 		</view>
@@ -91,10 +85,9 @@
 					start_time: [],
 				},
 				options: {
+					pageSize: 20,
 					pageCurrent: 1, // 当前页
 					total: 0, // 数据总量
-					pageSizeIndex: 0, // 与 pageSizeRange 一起计算得出 pageSize
-					pageSizeRange: [10, 20, 50, 100],
 				},
 				loading: false,
 				currentDateTab: 1,
@@ -108,13 +101,6 @@
 			}
 		},
 		computed: {
-			pageSize() {
-				const {
-					pageSizeRange,
-					pageSizeIndex
-				} = this.options
-				return pageSizeRange[pageSizeIndex]
-			},
 			chartTabs() {
 				const tabs = []
 				fieldsMap.forEach(item => {
@@ -216,7 +202,7 @@
 			changePlatform(id, index, name, item) {
 				this.getChannelData(null, id)
 				this.query.version_id = 0
-				console.log('-- item.code',item.code);
+				console.log('-- item.code', item.code);
 				this.query.uni_platform = item.code
 			},
 			changeTimeRange(id, index) {
@@ -236,12 +222,9 @@
 				this.getTabelData(this.query)
 			},
 
-			changePageSize(e) {
-				const {
-					value
-				} = e.detail
+			changePageSize(pageSize) {
+				this.options.pageSize = pageSize
 				// this.options.pageCurrent = 1 // 重置分页
-				this.options.pageSizeIndex = value
 				this.getTabelData(this.query)
 			},
 
@@ -319,8 +302,8 @@
 					.groupBy('start_time')
 					.groupField(stringifyGroupField(fieldsMap))
 					.orderBy('start_time', 'desc')
-					.skip((pageCurrent - 1) * this.pageSize)
-					.limit(this.pageSize)
+					.skip((pageCurrent - 1) * this.options.pageSize)
+					.limit(this.options.pageSize)
 					.get({
 						getCount: true
 					})